--- pl_bolts/models/reinforcement/__init__.py
@@ -1,8 +1,0 @@
-from pl_bolts.models.reinforcement.double_dqn.model import DoubleDQN
-from pl_bolts.models.reinforcement.dqn.model import DQN
-from pl_bolts.models.reinforcement.dueling_dqn.model import DuelingDQN
-from pl_bolts.models.reinforcement.n_step_dqn.model import NStepDQN
-from pl_bolts.models.reinforcement.noisy_dqn.model import NoisyDQN
-from pl_bolts.models.reinforcement.per_dqn.model import PERDQN
-from pl_bolts.models.reinforcement.reinforce.model import Reinforce
-from pl_bolts.models.reinforcement.vanilla_policy_gradient.model import PolicyGradient
