+++ pl_bolts/models/rl/__init__.py
@@ -1,0 +1,8 @@
+from pl_bolts.models.rl.double_dqn.model import DoubleDQN
+from pl_bolts.models.rl.dqn.model import DQN
+from pl_bolts.models.rl.dueling_dqn.model import DuelingDQN
+from pl_bolts.models.rl.n_step_dqn.model import NStepDQN
+from pl_bolts.models.rl.noisy_dqn.model import NoisyDQN
+from pl_bolts.models.rl.per_dqn.model import PERDQN
+from pl_bolts.models.rl.reinforce.model import Reinforce
+from pl_bolts.models.rl.vanilla_policy_gradient.model import PolicyGradient
