+++ pl_bolts/models/rl/n_step_dqn/model.py
@@ -3,8 +3,8 @@ N Step Deep Q-network
 """
 import torch
 
-from pl_bolts.models.reinforcement.common.experience import NStepExperienceSource
-from pl_bolts.models.reinforcement.dqn.model import DQN
+from pl_bolts.models.rl.common.experience import NStepExperienceSource
+from pl_bolts.models.rl.dqn.model import DQN
 
 
 class NStepDQN(DQN):
@@ -37,7 +37,7 @@ class NStepDQN(DQN):
 
         Example:
 
-            >>> from pl_bolts.models.reinforcement.n_step_dqn.model import NStepDQN
+            >>> from pl_bolts.models.rl.n_step_dqn.model import NStepDQN
             ...
             >>> model = NStepDQN("PongNoFrameskip-v4")
