+++ tests/models/test_reinforcement/unit/test_vpg.py
@@ -7,12 +7,12 @@ import numpy as np
 import torch
 from torch.utils.data import DataLoader
 
-from pl_bolts.models.reinforcement.common import cli
-from pl_bolts.models.reinforcement.common.agents import Agent
-from pl_bolts.models.reinforcement.common.experience import EpisodicExperienceStream
-from pl_bolts.models.reinforcement.common.networks import MLP
-from pl_bolts.models.reinforcement.common.wrappers import ToTensor
-from pl_bolts.models.reinforcement.vanilla_policy_gradient.model import PolicyGradient
+from pl_bolts.models.rl.common import cli
+from pl_bolts.models.rl.common.agents import Agent
+from pl_bolts.models.rl.common.experience import EpisodicExperienceStream
+from pl_bolts.models.rl.common.networks import MLP
+from pl_bolts.models.rl.common.wrappers import ToTensor
+from pl_bolts.models.rl.vanilla_policy_gradient.model import PolicyGradient
 
 
 class TestPolicyGradient(TestCase):
